[320] refactor Complete Your Profile
