+++ .storybook/preview.js
@@ -1,3 +1,4 @@
+import "../src/css/main.css";
 /** @type { import('@storybook/react').Preview } */
 const preview = {
   parameters: {
