+++ src/js/common/stories/HowItWorksWizard.stories.js
@@ -1,9 +1,44 @@
 import React from 'react';
 import styled from 'styled-components';
-import HowItWorks from '../../components/CompleteYourProfile/CompleteYourProfile2024';
+import HowItWorks from '../../components/CompleteYourProfile/HowItWorksWizard';
+
+const steps = [
+  {
+    id: 1,
+    title: 'How WeVote works',
+    buttonText: '',
+    completed: true,
+    description: '',
+    onClick: '',
+    titleCanBeClicked: true,
+    width: '33.33%',
+  },
+  {
+    id: 2,
+    title: 'Your personalized score',
+    buttonText: '',
+    completed: false,
+    description: '',
+    onClick: '',
+    titleCanBeClicked: true,
+    width: '33.33%',
+  },
+  {
+    id: 3,
+    title: 'Sign in or join to save your ballot choices/settings',
+    buttonText: 'Sign up to save choices',
+    completed: false,
+    description: '',
+    onClick: '',
+    titleCanBeClicked: false,
+    width: '33.33%',
+  },
+];
+
+const activeStep = 1;
 
 export default {
-  title: 'Design System/Wizard',
+  title: 'Design System/CompleteYourProfile',
   component: HowItWorks,
   parameters: {
     layout: 'centered',
@@ -19,6 +54,6 @@ const Container = styled.div`
 
 export const HowItWorksWizard = () => (
   <Container>
-    <HowItWorks />
+    <HowItWorks steps={steps} activeStep={activeStep} />
   </Container>
 );
